Marking this spot: drawing graphs
First step of refactoring: Moving all the graphing code out of filegallery.js into graph.js. Now it draw the graph again, but nothing’s clickable yet. ENCD-3510
