+++ src/encoded/static/components/filegallery.js
@@ -2,21 +2,18 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import _ from 'underscore';
 import moment from 'moment';
-import * as globals from './globals';
 import { Panel, PanelHeading, TabPanel, TabPanelPane } from '../libs/bootstrap/panel';
 import { Modal, ModalHeader, ModalBody, ModalFooter } from '../libs/bootstrap/modal';
+import { collapseIcon } from '../libs/svg-icons';
 import { auditDecor, auditsDisplayed, AuditIcon } from './audit';
-import StatusLabel from './statuslabel';
+import { FetchedData, Param } from './fetched';
+import * as globals from './globals';
+import { FileGraph } from './graph';
 import { requestFiles, DownloadableAccession, BrowserSelector } from './objectutils';
-import { Graph, JsonGraph } from './graph';
 import { qcModalContent, qcIdToDisplay } from './quality_metric';
 import { softwareVersionList } from './software';
-import { FetchedData, Param } from './fetched';
-import { collapseIcon } from '../libs/svg-icons';
 import { SortTablePanel, SortTable } from './sorttable';
-
-
-const MINIMUM_COALESCE_COUNT = 5; // Minimum number of files in a coalescing group
+import StatusLabel from './statuslabel';
 
 
 // Get the audit icon for the highest audit level in the given file.
@@ -901,477 +898,6 @@ function collectAssembliesAnnotations(files) {
 }
 
 
-// Handle graphing throws. Exported for Jest tests.
-export function GraphException(message, file0, file1) {
-    this.message = message;
-    if (file0) {
-        this.file0 = file0;
-    }
-    if (file1) {
-        this.file1 = file1;
-    }
-}
-
-
-// Map a QC object to its corresponding two-letter abbreviation for the graph.
-function qcAbbr(qc) {
-    // As we add more QC object types, add to this object.
-    const qcAbbrMap = {
-        BigwigcorrelateQualityMetric: 'BC',
-        BismarkQualityMetric: 'BK',
-        ChipSeqFilterQualityMetric: 'CF',
-        ComplexityXcorrQualityMetric: 'CX',
-        CorrelationQualityMetric: 'CN',
-        CpgCorrelationQualityMetric: 'CC',
-        DuplicatesQualityMetric: 'DS',
-        EdwbamstatsQualityMetric: 'EB',
-        EdwcomparepeaksQualityMetric: 'EP',
-        Encode2ChipSeqQualityMetric: 'EC',
-        FastqcQualityMetric: 'FQ',
-        FilteringQualityMetric: 'FG',
-        GenericQualityMetric: 'GN',
-        HotspotQualityMetric: 'HS',
-        IDRQualityMetric: 'ID',
-        IdrSummaryQualityMetric: 'IS',
-        MadQualityMetric: 'MD',
-        SamtoolsFlagstatsQualityMetric: 'SF',
-        SamtoolsStatsQualityMetric: 'SS',
-        StarQualityMetric: 'SR',
-        TrimmingQualityMetric: 'TG',
-    };
-
-    let abbr = qcAbbrMap[qc['@type'][0]];
-    if (!abbr) {
-        // 'QC' is the generic, unmatched abbreviation if qcAbbrMap doesn't have a match.
-        abbr = 'QC';
-    }
-    return abbr;
-}
-
-
-export function assembleGraph(context, session, infoNodeId, files, filterAssembly, filterAnnotation, colorize) {
-    // Calculate a step ID from a file's derived_from array.
-    function rDerivedFileIds(file) {
-        if (file.derived_from && file.derived_from.length) {
-            return file.derived_from.sort().join();
-        }
-        return '';
-    }
-
-    // Calculate a QC node ID.
-    function rGenQcId(metric, file) {
-        return `qc:${metric['@id'] + file['@id']}`;
-    }
-
-    /**
-     * Generate a string of CSS classes for a file node. Plass the result into a `className` property of a component.
-     *
-     * @param {object-required} file - File we're generating the statuses for.
-     * @param {bool} active - True if the file is active and should be highlighted as such.
-     * @param (bool) colorize - True to colorize the nodes according to their status by adding a CSS class for their status
-     * @param {string} addClasses - CSS classes to add in addition to the ones generated by the file statuses.
-     */
-    function fileCssClassGen(file, active, colorizeNode, addClasses) {
-        let statusClass;
-        if (colorizeNode) {
-            statusClass = file.status.replace(/ /g, '-');
-        }
-        return `pipeline-node-file${active ? ' active' : ''}${colorizeNode ? ` ${statusClass}` : ''}${addClasses ? ` ${addClasses}` : ''}`;
-    }
-
-    const derivedFileIds = _.memoize(rDerivedFileIds, file => file['@id']);
-    const genQcId = _.memoize(rGenQcId, (metric, file) => metric['@id'] + file['@id']);
-
-    // Begin collecting up information about the files from the search result, and gathering their
-    // QC and analysis pipeline information.
-    const graphedFiles = {}; // All files in the graph, so table can link to it.
-    const allFiles = {}; // All searched files, keyed by file @id
-    let matchingFiles = {}; // All files that match the current assembly/annotation, keyed by file @id
-    const fileQcMetrics = {}; // List of all file QC metrics indexed by file @id
-    const allPipelines = {}; // List of all pipelines indexed by step @id
-    files.forEach((file) => {
-        // allFiles gets all files from search regardless of filtering.
-        allFiles[file['@id']] = file;
-
-        // matchingFiles gets just the files matching the given filtering assembly/annotation.
-        // Note that if all assemblies and annotations are selected, this function isn't called
-        // because no graph gets displayed in that case.
-        if ((file.assembly === filterAssembly) && ((!file.genome_annotation && !filterAnnotation) || (file.genome_annotation === filterAnnotation))) {
-            // Note whether any files have an analysis step
-            const fileAnalysisStep = file.analysis_step_version && file.analysis_step_version.analysis_step;
-            if (!fileAnalysisStep || (file.derived_from && file.derived_from.length)) {
-                // File has no analysis step or derives from other files, so it can be included in
-                // the graph.
-                matchingFiles[file['@id']] = file;
-
-                // Collect any QC info that applies to this file and make it searchable by file
-                // @id.
-                if (file.quality_metrics && file.quality_metrics.length) {
-                    fileQcMetrics[file['@id']] = file.quality_metrics;
-                }
-
-                // Save the pipeline array used for each step used by the file.
-                if (fileAnalysisStep) {
-                    allPipelines[fileAnalysisStep['@id']] = fileAnalysisStep.pipelines;
-                }
-            } // else file has analysis step but no derived from -- can't include in graph.
-        }
-    });
-
-    // Generate a list of file @ids that other files (matching the current assembly and annotation)
-    // derive from (i.e. files referenced in other files' derived_from). allDerivedFroms is keyed
-    // by the derived-from file @id (whether it matches the current assembly and annotation or not)
-    // and has an array of all files that derive from it for its value. So for example:
-    //
-    // allDerivedFroms = {
-    //     /files/<matching accession>: [matching file, matching file],
-    //     /files/<contributing accession>: [matching file, matching file],
-    //     /files/<missing accession>: [matching file, matching file],
-    // }
-    const allDerivedFroms = {};
-    Object.keys(matchingFiles).forEach((matchingFileId) => {
-        const matchingFile = matchingFiles[matchingFileId];
-        if (matchingFile.derived_from && matchingFile.derived_from.length) {
-            matchingFile.derived_from.forEach((derivedFromAtId) => {
-                // Copy reference to allFiles copy of file. Will be undefined for missing and
-                // contributing files.
-                if (allDerivedFroms[derivedFromAtId]) {
-                    // Already saw a file derive from this one, so add the new reference to the end
-                    // of the array of derived-from files.
-                    allDerivedFroms[derivedFromAtId].push(matchingFile);
-                } else {
-                    // Never saw a file derive from this one, so make a new array with a reference
-                    // to it.
-                    allDerivedFroms[derivedFromAtId] = [matchingFile];
-                }
-            });
-        }
-    });
-    // Remember, at this stage allDerivedFroms includes keys for missing files, files not matching
-    // the chosen assembly/annotation, and contributing files.
-
-    // Filter any "island" files out of matchingFiles -- that is, files that derive from no other
-    // files, and no other files derive from it.
-    matchingFiles = (function matchingFilesFunc() {
-        const noIslandFiles = {};
-        Object.keys(matchingFiles).forEach((matchingFileId) => {
-            const matchingFile = matchingFiles[matchingFileId];
-            if ((matchingFile.derived_from && matchingFile.derived_from.length) || allDerivedFroms[matchingFileId]) {
-                // This file either has derived_from set, or other files derive from it. Copy it to
-                // our destination object.
-                noIslandFiles[matchingFileId] = matchingFile;
-            }
-        });
-        return noIslandFiles;
-    }());
-    if (Object.keys(matchingFiles).length === 0) {
-        throw new GraphException('No graph: no file relationships for the selected assembly/annotation');
-    }
-    // At this stage, any files in matchingFiles will be rendered. We just have to figure out what
-    // other files need rendering, like raw sequencing files, contributing files, and derived-from
-    // files that have a non-matching annotation and assembly.
-
-    const allReplicates = {}; // All file's replicates as keys; each key references an array of files
-    Object.keys(matchingFiles).forEach((matchingFileId) => {
-        // If the file is part of a single biological replicate, add it to an array of files, where
-        // the arrays are in an object keyed by their relevant biological replicate number.
-        const matchingFile = matchingFiles[matchingFileId];
-        let replicateNum = (matchingFile.biological_replicates && matchingFile.biological_replicates.length === 1) ? matchingFile.biological_replicates[0] : undefined;
-        if (replicateNum) {
-            if (allReplicates[replicateNum]) {
-                allReplicates[replicateNum].push(matchingFile);
-            } else {
-                allReplicates[replicateNum] = [matchingFile];
-            }
-        }
-
-        // Add each file that a matching file derives from to the replicates.
-        if (matchingFile.derived_from && matchingFile.derived_from.length) {
-            matchingFile.derived_from.forEach((derivedFromAtId) => {
-                const file = allFiles[derivedFromAtId];
-                if (file) {
-                    replicateNum = (file.biological_replicates && file.biological_replicates.length === 1) ? file.biological_replicates[0] : undefined;
-                    if (replicateNum) {
-                        if (allReplicates[replicateNum]) {
-                            allReplicates[replicateNum].push(matchingFile);
-                        } else {
-                            allReplicates[replicateNum] = [matchingFile];
-                        }
-                    }
-                }
-            });
-        }
-    });
-
-    // Make a list of contributing files that matchingFiles files derive from.
-    const usedContributingFiles = {};
-    if (context.contributing_files && context.contributing_files.length) {
-        context.contributing_files.forEach((contributingFileAtId) => {
-            if (contributingFileAtId in allDerivedFroms) {
-                usedContributingFiles[contributingFileAtId] = allDerivedFroms[contributingFileAtId];
-            }
-        });
-    }
-
-    // Go through each used contributing file and set a property within it showing which files
-    // derive from it. We'll need that for coalescing contributing files.
-    const allCoalesced = {};
-    let coalescingGroups = {};
-    if (Object.keys(usedContributingFiles).length) {
-        // Now use the derivedFiles property of every contributing file to group them into potential
-        // coalescing nodes. `coalescingGroups` gets assigned an object keyed by dataset file ids
-        // hashed to a stringified 32-bit integer, and mapped to an array of contributing files they
-        // derive from.
-        coalescingGroups = _(Object.keys(usedContributingFiles)).groupBy((contributingFileAtId) => {
-            const derivedFiles = usedContributingFiles[contributingFileAtId];
-            return globals.hashCode(derivedFiles.map(derivedFile => derivedFile['@id']).join(',')).toString();
-        });
-
-        // Set a `coalescingGroup` property in each contributing file with its coalescing group's hash
-        // value. That'll be important when we add step nodes.
-        const coalescingGroupKeys = Object.keys(coalescingGroups);
-        if (coalescingGroupKeys && coalescingGroupKeys.length) {
-            coalescingGroupKeys.forEach((groupHash) => {
-                const group = coalescingGroups[groupHash];
-                if (group.length >= MINIMUM_COALESCE_COUNT) {
-                    // Number of files in the coalescing group is at least the minimum number of files we
-                    // allow in a coalescing group. Mark every contributing file in the group with the
-                    // group's hash value in a `coalescingGroup` property that step node can connnect to.
-                    group.forEach((contributingFileAtId) => {
-                        allCoalesced[contributingFileAtId] = groupHash;
-
-                        // Remove coalesced files from usedContributingFiles because we don't want
-                        // to render individual files that have been coalesced.
-                        delete usedContributingFiles[contributingFileAtId];
-                    });
-                } else {
-                    // The number of contributing files in a coalescing group isn't above our
-                    // threshold. Don't use this coalescingGroup anymore and just render them the
-                    // same as normal files.
-                    delete coalescingGroups[groupHash];
-                }
-            });
-        }
-    }
-
-    // See if we have any derived_from files that we have no information on, likely because they're
-    // not released and we're not logged in. We'll render them with information-less dummy nodes.
-    const allMissingFiles = [];
-    Object.keys(allDerivedFroms).forEach((derivedFromFileAtId) => {
-        if (!allFiles[derivedFromFileAtId] && !allCoalesced[derivedFromFileAtId]) {
-            // The derived-from file isn't in our dataset file list, nor in coalesced contributing
-            // files. Now see if it's in non-coalesced contributing files.
-            if (!usedContributingFiles[derivedFromFileAtId]) {
-                allMissingFiles.push(derivedFromFileAtId);
-            }
-        }
-    });
-
-    // Create an empty graph architecture that we fill in next.
-    const jsonGraph = new JsonGraph(context.accession);
-
-    // Create nodes for the replicates.
-    Object.keys(allReplicates).forEach((replicateNum) => {
-        if (allReplicates[replicateNum] && allReplicates[replicateNum].length) {
-            jsonGraph.addNode(`rep:${replicateNum}`, `Replicate ${replicateNum}`, {
-                cssClass: 'pipeline-replicate',
-                type: 'Rep',
-                shape: 'rect',
-                cornerRadius: 0,
-            });
-        }
-    });
-
-    // Go through each file matching the currently selected assembly/annotation and add it to our
-    // graph.
-    Object.keys(matchingFiles).forEach((fileId) => {
-        const file = matchingFiles[fileId];
-        const fileNodeId = `file:${file['@id']}`;
-        const fileNodeLabel = `${file.title} (${file.output_type})`;
-        const fileCssClass = fileCssClassGen(file, infoNodeId === fileNodeId, colorize);
-        const fileRef = file;
-        const replicateNode = (file.biological_replicates && file.biological_replicates.length === 1) ? jsonGraph.getNode(`rep:${file.biological_replicates[0]}`) : null;
-        let metricsInfo;
-
-        // Add QC metrics info from the file to the list to generate the nodes later.
-        if (fileQcMetrics[fileId] && fileQcMetrics[fileId].length) {
-            const sortedMetrics = fileQcMetrics[fileId].sort((a, b) => (a['@type'][0] > b['@type'][0] ? 1 : (a['@type'][0] < b['@type'][0] ? -1 : 0)));
-            metricsInfo = sortedMetrics.map((metric) => {
-                const qcId = genQcId(metric, file);
-                return {
-                    id: qcId,
-                    label: qcAbbr(metric),
-                    '@type': ['QualityMetric'],
-                    class: `pipeline-node-qc-metric${infoNodeId === qcId ? ' active' : ''}`,
-                    tooltip: true,
-                    ref: metric,
-                    parent: file,
-                };
-            });
-        }
-
-        // Add a node for a regular searched file.
-        jsonGraph.addNode(fileNodeId, fileNodeLabel, {
-            cssClass: fileCssClass,
-            type: 'File',
-            shape: 'rect',
-            cornerRadius: 16,
-            parentNode: replicateNode,
-            ref: fileRef,
-        }, metricsInfo);
-
-        // Add the matching file to our list of "all" graphed files.
-        graphedFiles[fileId] = file;
-
-        // Figure out the analysis step we need to render between the node we just rendered and its
-        // derived_from.
-        let stepId;
-        let label;
-        let pipelineInfo;
-        let error;
-        const fileAnalysisStep = file.analysis_step_version && file.analysis_step_version.analysis_step;
-        if (fileAnalysisStep) {
-            // Make an ID and label for the step
-            stepId = `step:${derivedFileIds(file) + fileAnalysisStep['@id']}`;
-            label = fileAnalysisStep.analysis_step_types;
-            pipelineInfo = allPipelines[fileAnalysisStep['@id']];
-            error = false;
-        } else if (derivedFileIds(file)) {
-            // File derives from others, but no analysis step; make dummy step.
-            stepId = `error:${derivedFileIds(file)}`;
-            label = 'Software unknown';
-            pipelineInfo = null;
-            error = true;
-        } else {
-            // No analysis step and no derived_from; don't add a step.
-            stepId = '';
-        }
-
-        // If we have a step to render, do that here.
-        if (stepId) {
-            // Add the step to the graph only if we haven't for this derived-from set already
-            if (!jsonGraph.getNode(stepId)) {
-                jsonGraph.addNode(stepId, label, {
-                    cssClass: `pipeline-node-analysis-step${(infoNodeId === stepId ? ' active' : '') + (error ? ' error' : '')}`,
-                    type: 'Step',
-                    shape: 'rect',
-                    cornerRadius: 4,
-                    parentNode: replicateNode,
-                    ref: fileAnalysisStep,
-                    pipelines: pipelineInfo,
-                    fileId: file['@id'],
-                    fileAccession: file.title,
-                    stepVersion: file.analysis_step_version,
-                });
-            }
-
-            // Connect the file to the step, and the step to the derived_from files
-            jsonGraph.addEdge(stepId, fileNodeId);
-            file.derived_from.forEach((derivedFromAtId) => {
-                const derivedFromFile = allFiles[derivedFromAtId] || allMissingFiles.some(missingFileId => missingFileId === derivedFromAtId);
-                if (derivedFromFile) {
-                    // Not derived from a contributing file; just add edges normally.
-                    const derivedFileId = `file:${derivedFromAtId}`;
-                    if (!jsonGraph.getEdge(derivedFileId, stepId)) {
-                        jsonGraph.addEdge(derivedFileId, stepId);
-                    }
-                } else {
-                    // File derived from a contributing file; add edges to a coalesced node
-                    // that we'll add to the graph later.
-                    const coalescedContributing = allCoalesced[derivedFromAtId];
-                    if (coalescedContributing) {
-                        // Rendering a coalesced contributing file.
-                        const derivedFileId = `coalesced:${coalescedContributing}`;
-                        if (!jsonGraph.getEdge(derivedFileId, stepId)) {
-                            jsonGraph.addEdge(derivedFileId, stepId);
-                        }
-                    } else if (usedContributingFiles[derivedFromAtId]) {
-                        const derivedFileId = `file:${derivedFromAtId}`;
-                        if (!jsonGraph.getEdge(derivedFileId, stepId)) {
-                            jsonGraph.addEdge(derivedFileId, stepId);
-                        }
-                    }
-                }
-            });
-        }
-    });
-
-    // Go through each derived-from file and add it to our graph.
-    Object.keys(allDerivedFroms).forEach((fileId) => {
-        const file = allFiles[fileId];
-        if (file && !matchingFiles[fileId]) {
-            const fileNodeId = `file:${file['@id']}`;
-            const fileNodeLabel = `${file.title} (${file.output_type})`;
-            const fileCssClass = fileCssClassGen(file, infoNodeId === fileNodeId, colorize);
-            const fileRef = file;
-            const replicateNode = (file.biological_replicates && file.biological_replicates.length === 1) ? jsonGraph.getNode(`rep:${file.biological_replicates[0]}`) : null;
-
-            jsonGraph.addNode(fileNodeId, fileNodeLabel, {
-                cssClass: fileCssClass,
-                type: 'File',
-                shape: 'rect',
-                cornerRadius: 16,
-                parentNode: replicateNode,
-                ref: fileRef,
-            });
-
-            // Add the derived-from file to our list of "all" graphed files.
-            graphedFiles[fileId] = file;
-        }
-    });
-
-    // Go through each derived-from contributing file and add it to our graph.
-    Object.keys(usedContributingFiles).forEach((fileAtId) => {
-        const fileNodeId = `file:${fileAtId}`;
-        const fileNodeLabel = `${globals.atIdToAccession(fileAtId)}`;
-        const fileCssClass = `pipeline-node-file contributing${infoNodeId === fileNodeId ? ' active' : ''}`;
-
-        jsonGraph.addNode(fileNodeId, fileNodeLabel, {
-            cssClass: fileCssClass,
-            type: 'File',
-            shape: 'rect',
-            cornerRadius: 16,
-            contributing: fileAtId,
-            ref: {},
-        });
-    });
-
-    // Now add coalesced nodes to the graph.
-    Object.keys(coalescingGroups).forEach((groupHash) => {
-        const coalescingGroup = coalescingGroups[groupHash];
-        if (coalescingGroup.length) {
-            const fileNodeId = `coalesced:${groupHash}`;
-            const fileCssClass = `pipeline-node-file contributing${infoNodeId === fileNodeId ? ' active' : ''}`;
-            jsonGraph.addNode(fileNodeId, `${coalescingGroup.length} contributing files`, {
-                cssClass: fileCssClass,
-                type: 'File',
-                shape: 'stack',
-                cornerRadius: 16,
-                contributing: groupHash,
-                ref: coalescingGroup,
-            });
-        }
-    });
-
-    // Add missing-file nodes to the graph.
-    allMissingFiles.forEach((missingFileAtId) => {
-        const fileNodeAccession = globals.atIdToAccession(missingFileAtId);
-        const fileNodeId = `file:${missingFileAtId}`;
-        const fileNodeLabel = `${fileNodeAccession} (unknown)`;
-        const fileCssClass = 'pipeline-node-file error';
-
-        jsonGraph.addNode(fileNodeId, fileNodeLabel, {
-            cssClass: fileCssClass,
-            type: 'File',
-            shape: 'rect',
-            cornerRadius: 16,
-        });
-    });
-
-    return { graph: jsonGraph, graphedFiles };
-}
-
-
 export function qcDetailsView(metrics, schemas) {
     const qc = metrics.ref;
 
@@ -1611,7 +1137,8 @@ class FileGalleryRendererComponent extends React.Component {
         return files;
     }
 
-    // Handle a click in a graph node
+    // Handle a click in a graph node. This also handles clicks on the info button of files in the
+    // file table.
     handleNodeClick(nodeId) {
         this.setInfoNodeId(nodeId);
         this.setInfoNodeVisible(true);
@@ -1707,9 +1234,7 @@ class FileGalleryRendererComponent extends React.Component {
         const { context, data, schemas, hideGraph } = this.props;
         let selectedAssembly = '';
         let selectedAnnotation = '';
-        let jsonGraph;
         let allGraphedFiles;
-        let meta = null;
         const files = (data ? data['@graph'] : []).concat(this.state.relatedFiles); // Array of searched files arrives in data.@graph result
         if (files.length === 0) {
             return null;
@@ -1723,36 +1248,20 @@ class FileGalleryRendererComponent extends React.Component {
         }
 
         // Get a list of files for the graph (filters out excluded files if requested by the user).
-        const graphFiles = this.filterForInclusion(files);
-
-        // Build node graph of the files and analysis steps with this experiment
-        if (graphFiles && graphFiles.length && !hideGraph) {
-            try {
-                const { graph, graphedFiles } = assembleGraph(context, this.context.session, this.state.infoNodeId, graphFiles, selectedAssembly, selectedAnnotation, this.state.inclusionOn);
-                jsonGraph = graph;
-                allGraphedFiles = (selectedAssembly || selectedAnnotation) ? graphedFiles : {};
-                meta = this.detailNodes(jsonGraph, this.state.infoNodeId, this.context.session, this.context.session_properties);
-            } catch (e) {
-                jsonGraph = null;
-                allGraphedFiles = {};
-                console.warn(e.message + (e.file0 ? ` -- file0:${e.file0}` : '') + (e.file1 ? ` -- file1:${e.file1}` : ''));
-            }
-        }
+        const includedFiles = this.filterForInclusion(files);
 
         // Prepare to display the file information modal.
-        const modalTypeMap = {
-            File: 'file',
-            Step: 'analysis-step',
-            QualityMetric: 'quality-metric',
-        };
-        const modalClass = meta ? `graph-modal-${modalTypeMap[meta.type]}` : '';
+        // const modalTypeMap = {
+        //     File: 'file',
+        //     Step: 'analysis-step',
+        //     QualityMetric: 'quality-metric',
+        // };
+        // const modalClass = meta ? `graph-modal-${modalTypeMap[meta.type]}` : '';
 
-        // Generate the file table first so we can render it the same way regardless of whether
-        // we have a file graph or not.
         const fileTable = (
             <FileTable
                 {...this.props}
-                items={graphFiles}
+                items={includedFiles}
                 selectedFilterValue={this.state.selectedFilterValue}
                 filterOptions={filterOptions}
                 graphedFiles={allGraphedFiles}
@@ -1794,16 +1303,11 @@ class FileGalleryRendererComponent extends React.Component {
                         <TabPanelPane key="graph">
                             {!hideGraph ?
                                 <FileGraph
-                                    context={context}
-                                    items={graphFiles}
-                                    graph={jsonGraph}
+                                    dataset={context}
+                                    files={includedFiles}
                                     selectedAssembly={selectedAssembly}
                                     selectedAnnotation={selectedAnnotation}
                                     handleNodeClick={this.handleNodeClick}
-                                    setInfoNodeId={this.setInfoNodeId}
-                                    setInfoNodeVisible={this.setInfoNodeVisible}
-                                    schemas={schemas}
-                                    forceRedraw
                                 />
                             : null}
                         </TabPanelPane>
@@ -1818,9 +1322,9 @@ class FileGalleryRendererComponent extends React.Component {
                 :
                     <div>{fileTable}</div>
                 }
-                {meta && this.state.infoNodeVisible ?
+                {/* {meta && this.state.infoNodeVisible ?
                     <Modal closeModal={this.closeModal}>
-                        <ModalHeader closeModal={this.closeModal} addCss={modalClass}>
+                        <ModalHeader closeModal={this.closeModal}>
                             {meta ? meta.header : null}
                         </ModalHeader>
                         <ModalBody>
@@ -1828,7 +1332,7 @@ class FileGalleryRendererComponent extends React.Component {
                         </ModalBody>
                         <ModalFooter closeModal={<button className="btn btn-info" onClick={this.closeModal}>Close</button>} />
                     </Modal>
-                : null}
+                : null} */}
             </Panel>
         );
     }
@@ -1899,47 +1403,6 @@ FilterMenu.propTypes = {
     selectedFilterValue: PropTypes.string, // Currently selected filter
     filterOptions: PropTypes.array.isRequired, // Contents of the filtering menu
     handleFilterChange: PropTypes.func.isRequired, // Call when a filtering option changes
-};
-
-
-class FileGraph extends React.Component {
-    constructor() {
-        super();
-
-        // Initialize React state variables.
-        this.state = {
-            contributingFiles: {}, // List of contributing file objects we've requested; acts as a cache too
-            coalescedFiles: {}, // List of coalesced files we've requested; acts as a cache too
-            infoModalOpen: false, // Graph information modal open
-        };
-    }
-
-    render() {
-        const { items, graph, selectedAssembly, selectedAnnotation, handleNodeClick } = this.props;
-        const files = items;
-
-        // Build node graph of the files and analysis steps with this experiment
-        if (files && files.length) {
-            // If we have a graph, or if we have a selected assembly/annotation, draw the graph panel
-            const goodGraph = graph && Object.keys(graph).length;
-            if (goodGraph) {
-                if (selectedAssembly || selectedAnnotation) {
-                    return <Graph graph={graph} nodeClickHandler={handleNodeClick} nodeMouseenterHandler={this.handleHoverIn} nodeMouseleaveHandler={this.handleHoverOut} noDefaultClasses forceRedraw />;
-                }
-                return <p className="browser-error">Choose an assembly to see file association graph</p>;
-            }
-            return <p className="browser-error">Graph not applicable for the selected assembly/annotation.</p>;
-        }
-        return <p className="browser-error">Graph not applicable.</p>;
-    }
-}
-
-FileGraph.propTypes = {
-    items: PropTypes.array, // Array of files we're graphing
-    graph: PropTypes.object, // JsonGraph object generated from files
-    selectedAssembly: PropTypes.string, // Currently selected assembly
-    selectedAnnotation: PropTypes.string, // Currently selected annotation
-    handleNodeClick: PropTypes.func, // Parent function to call when a graph node is clicked
 };
